Merge Sort Part 2 checkpoint
Added part 2 of mergesort, which builds on the third part of the recursion series as well as the first part of mergesort.

Added an expostion framework that seems to help keep the simulated code cleaner. The first use of it, in scene 2, is made more complex by the need to hide the code half way through the scene.